continue later of the travel

+++ app/Http/Controllers/AdminUsersController.php
@@ -161,7 +161,12 @@ class AdminUsersController extends Controller
     public function destroy($id)
     {
         $user = User::findOrFail($id);
-        unlink(public_path().$user->photo->file);
+        if($user->photo_id!=NULL){
+            unlink(public_path().$user->photo->file);
+            Photo::findOrFail($user->photo_id)->delete();
+        }
+
+
